fix: use browser tools by writing Python code, added stripping of data:image's
